Tailorized "2005-02-15 03:41:20 by logg" Added solver base class and cleaned up Poisson module and main.cpp. The base class is just a simple way of getting all the includes for a module. (Should be simple for users to add new modules.)
Also modified the Poisson test problem somewhat and modified the
checkerror.m script for computing the error. To run the test, just do

    octave checkerror.m

Also added an option to GMRES solver to be silent and not report
the number of iterations each time (good for the ODE solver). I think
we need to figure out a way to avoid putting a bunch of setFoo()
functions into the GMRES solver. (Possible redesign of the settings
system.)

Remains to do:

 - Get vector-valued functions (taking extra argument i) working
 - Make output work for NewFunction and modify Poisson module
 - Modify configure script accordning to Garth. I think there
   needs to be an option like --with-petsc=dir in addition to
   checking PETSC_DIR. The default should be /usr/lib/petsc.

If we do this, I think it's enough for a new release. For the
next release, we need to do the following:

 - Move all other modules to new code
 - Rename all NewFoo to Foo

/Anders