Tailorized "2005-05-11 22:17:41 by logg" Updates for latest (CVS) version of FFC. DOLFIN now uses the map automatically created by FFC to map local dofs to global dofs.
Also changed the assembler to generate two dof maps, one for test and
one for trial space. We might want to think of adding an optimization
later in case test and trial spaces are the same.

Higher degree Lagrange elements now work in principle, but the projection
of user-defined functions onto higher-degree finite element spaces has
not been implemented. Hopefully we can have something working for this
tomorrow.

I temporarily disabled the compilation of the Navier-Stokes module (since
it takes so long to compile the forms). Should be much faster when the
stabilization is compiled with piecewise constant velocity.