Tailorized "2005-06-05 17:06:36 by logg" Updated MassMatrix and StiffnessMatrix for new system.
MassMatrix and StiffnessMatrix are easy-to-use shortcuts for linear
elements. Normally, assembly is through FFC, but sometimes it may
be convenient to just declare a mass or stiffness matrix for linear
elements.

For example, to create the standard stiffness matrix on the unit
square with homogeneous Neumann boundary conditions, just do

  UnitSquare mesh(16, 16); // Total number of triangles is 2*16*16
  StiffnessMatrix A(mesh); // A is a (PETSc) matrix