Tailorized "2005-06-30 03:08:38 by logg" A number of assembly updates:
 - Make FEM::assemble() estimate the number of nonzeros in each row
 - Implement Matrix::init(M, N, nzmax)
 - Add Matrix::nz(), Matrix::nzsum() and Matrix::nzmax()
 - Improve Mesh::disp()
 - Add FiniteElement::disp() and FEM::disp() (useful for debugging)
 - Remove old class SparseMatrix
 - Change FEM::setBC() --> FEM::applyBC()
 - Change Mesh::tetrahedrons --> Mesh::tetrahedra

Still working on getting the assembly working for higher-order
Lagrange elements. Seems to be working for general order elements
on triangles and for q <= 3 tetrahedra. For q >= 4, the alignment
of faces is needed and I still need to track down a bug somewhere.

Check the test program in src/demo/fem/convergence/ for current status.