Allow local data to be attached to Functions:
  void attach(Vector& x,              bool local = false);
  void attach(Mesh& mesh,             bool local = false);
  void attach(FiniteElement& element, bool local = false);

If local is set, the data will be considered local to the Function.
The Function takes responsibility for the data (including deletion at
the appropriate time). Use with caution.

This is mainly to be used when reading Function data from file, when we
need to create the Vector, the Mesh and the FiniteElement (they are not
references to somebody else's data).