Initial fixes for ODE benchmark:
Results are now similar (close enough) to the results with
DOLFIN 0.5.12. A few default parameters were changed in the
Krylov solvers in addition to moving to PETSc 2.3.1 which
affected the multi-adaptive Newton solver. Scaling the system
with the size of the residual before solving and rescaling
after in each Newton step fixes the problem. (Scaling only
done locally in ODE solvers, not in Krylov solver.)

Next step is to clean up benchmarks and generate results.