Add library of predefined finite elements. Elements and dof maps may now be generated from their signature. For example:
  ufc::finite_element* = ElementLibrary::create_finite_element("Lagrange finite element of degree 3 on a tetrahedron");

This will be used for input of Functions from XML files, where only
the signatures of the elements and dof maps are stored.

Which elements are available in the library can be chosen in
the script src/kernel/elements/generate.py.

This replaces the old FiniteElementSpec class which did something
similar.