Implement extraction of sub functions from a function. As before, we can once again do
   Function w;
   Function u = w[0];
   Function p = w[1];

where w is a Function representing the solution of a mixed system
(like Stokes). Now, both setting bcs for sub systems and extracting
functions as sub functions should work, but more testing is needed.

The extraction of a sub function copies the degrees of freedom to
a new (smaller size) vector but the mesh is shared.

Plotting the presure seems ok in Viper but plotting the velocity
field returns an error message (see viper-dev).

Also, exporting to .pvd gives segmentation fault.