Option for setting boundary conditions based on local geometrical search. (Patch from Kristian with some modifications.)
This makes it possible to set boundary conditions for elements which don't
(according to tabulate_facet_dofs) have any dofs on the boundary. This includes
most notably DG elements where all dofs are internal to the cell, so we
can now set also strong BCs for DG elements.

The local test for coordinate on facet is only implemented for 2D so far.
3D will come.