Several Python assembly optimizations and new options: The following options are now available:
  dolfin_set("optimize form", True)              # Use optimization -O2 when compiling generated code
  dolfin_set("optimize use dof map cache", True) # Store dof maps in cache for reuse
  dolfin_set("optimize use tensor cache", True)  # Store tensors in cache for reuse

  dolfin_set("optimize", True)                   # All of the above

Each one of these may make a significant difference for the
speed of assembly from Python, in particular for repeated
assembly.

With these options set, the difference in speed between assembly
in C++ and Python using the JIT compiler is negligible. (For forms
which don't depend on functions that are evaluated in Python.)