Add new utility function errornorm()
This function provides a simple way to compute the norm of the
difference between two given functions. It can be used to evaluate the
error of a given approximate solution when the exact solution is
known. In simple cases, one may just define

  e = u - uh

and evalute for example the square of the error in the L2 norm by

  e = u - uh
  assemble(e*e*dx, mesh)

However, this is not stable w.r.t. round-off errors considering that
the form compiler will expand the expression above to

  u*u*dx + uh*uh*dx - 2*u*uh*dx

and this might get further expanded into thousands of terms for
higher order elements. Thus, the error will be evaluated by adding
a large number of terms which should sum up to something close to
zero (if the error is small).

This function computes the error by first projecting both u and uh
to a common space (of high accuracy), then subtracting the two fields
(which is easy since they are expressed in the same basis) and then
evaluating the integral."""