Refactoring the slice handling of Vectors and Matrices in PyDOLFIN:   - Introduces a c++ wrapper class for different type of slices   - Implemented the c++ get and set functions as seperate functions     instead of extending the c++ proxy class   - All logic of which get/set function that should be called is kept     in the python proxy class.
In addition to the previous slice functionality for Vectors we can now
access single rows and columns in Matrices as Vectors:

  col_vector = A[0:-1:2,0]
  row_vector = A[0,[0,4,7,-1]]

and sub matrices as other matrices:

  inds = array([0,4,7,8,10])
  A_sub = A[inds,inds]

The sub matrix does only keep the non zero values.

Valid "slices" are Python slices, list and numpy arrays of integers

Setting single rows, and columns together with sub matrices are not yet
implemented for matrices.