Add simple code for base64 encoding of VTK output.
Code is commented out, but it does work. Option needs to be added for ascii/binary outoput.

base64 file size is not likely to be much smaller than ascii, but it can be compressed which will make a big difference.

(this was very painful thanks to the non-existent VTK documentation.)