Make Data::update public.
The Function design seems to be betting a bit complicated again. I just want a function in Coefficient that takes a point and a cell and evaluates the function.We have Coefficient::eval(double*, Data&), but the necessary functions to constract Data were private.

Also, if one calls Coefficient::evaluate(....), the coordinate are updated in the Data object (inside Coefficient::evaluate(.)), but not the cell.