Handle timings of sub tasks in benchmarks.
From README:

This script runs all executables/scripts named 'bench' found in the
directory tree and records the result for benchmark found in foo/bar
under the name 'foo-bar' in logs/bench.log.

Two pieces of information are extracted from each benchmark:

1. A short description of the benchmark. This is extracted as the
first line of output from the benchmark.

2. One or more timings. These are extracted from output in the form

  BENCH <totaltime>
  BENCH <subtask> <subtime>

If no output is given (or <totaltime> is not given), then the total
running time of the program is recorded.