Dramatic speedup of progress bar (and algorithms using it).
This speeds up the bench/mesh/unitcube benchmark by a factor 3.
This explains in part the regression noticed earlier for this
benchmark (and maybe others).

It should also improve the speed of assembly for simple forms.
Lets see what the benchbot says tomorrow.

The progress bar now tries to be intelligent by both checking the
time elapsed since the last update (so that the progress bar is
not updated too often) and iteratively trying to find out how
often it needs to check the current time (which is expensive).

Thanks to Johannes for noticing.