Temporarily use full sparsity pattern for all la backends to enable support for row versus column storage via the factory-returned sparsity pattern for a backend.
To make thing simple (esp. in parallel) it's useful that all backends have a sparsity pattern, but will soon make the distiction between full and reduced (just the total size and local range). This is effectively already the case is parallel where a sparsity pattern with just the local range data is used for STLMatrix.