Implement Test/TrialFunction differently in the PyDOLFIN layer.
Test and trial functions are created on first request and attached
to the FunctionSpaceBase object such that they can be reused when
requested later. This way we can use running counts and still make
sure the ordering of test/trial is fixed, thereby avoiding
an issue where TestFunction(V1) == TestFunction(V2) from an ufl
point of view if only the meshes in V1 and V2 differ.

Fixes
https://bugs.launchpad.net/dolfin/+bug/1083720