Insert integrals by cell if there are no internal facet integrals in the bilinear form.
The macro-element for dg mass matrices is larger than the sparsity pattern of the matrix, therefore if there are no internal facet
integrals switch back to cell wise matrix insertion.  This will modify the application of boundary conditions a bit, potentially
increasing the number of times a bc value is added to the rhs (and a 1 is added to the diagonal of the matrix).  If rescaled the
effect will also be slightly different though also shouldn't affect the end result.