Remove facet_orientation concept and Facet::adjacent_cells(facet_orientation).
Interior integrals will now determine which cell is + or - depending
on the natural ordering of the cell markers, i.e. such that the cell
marker for the minus side is smaller than the plus side.  If both cell
markers are the same, the ordering is arbitrary, determined by the
computed cell connectivity.