Add easy to use Python wrapper for LocalAssembler
The wrapper is not super efficient as it must create and destroy some
ufc and other helper objects on each call. The focus is on ease of use
for testing and prototyping. On larger problems looping over elements in
Python will never be a good way to get performance anyway, so this use
case is not intended and global assembly routines or the LocalSolver
should be used instead depending on what is needed.