Special treatment of global dofs in sparsity pattern
Track global dofs separately when building pattern with the assumption
that global dof contributes to the full row. This produces suboptimally
full rows for reals which are localized to only part of the mesh. But
this is only really fast solution. Anything else needs lookups in dense
rows which is now N^2 and could be improved to N*log(N) in principle.

This patch should improve complexity of sparsity pattern builder on real
spaces to linear.

(cherry picked from commit bc504dd4bfb5498b55ad14b4487d964561fa0b13)