Problem in this case in multimesh-volume/main.cpp (check log file lines 2075 and forwards). Particularly in the inc-exc of T_0^1 and T_3^1 we get at stage 1 at line 2272:
try intersect:
drawline([0.2875,0.7125],[0.2875,0.3875],'r',1,1,15);drawline([0.2875,0.3125],[0.2875,0.3375],'b',1,1,15);axis equal;

without detecting an intersection, This is a problem in collides_segment_segment_2d.

But: This is not the only problem. The more overarching problem is that it is more stable for the intersection inc-exc to use more volume intersection as we did before when the intersection inc-exc had its own implementation. We should probably go back to the specialized inc-exc.