Clean-up in Python division operators
Still not sure whether

    __truediv__ = lambda self, value: self.__div__(value)

pattern is correct. Work as a workaround in SWIG 3.0.8. Problem is that
the patern

%extend dolfin:Klass {
%pythoncode %{
    __truediv__ = __div__
%}
}

pattern does not work (in 3.0.8) because __div__ is injected later,
after the class is defined. On the other hand pattern

%pythoncode %{
   Klass.__truediv__ = Klass.__div__
%}

sometimes works, sometimes not. Don't know how to control its placement
in the generated code. It might be placed before

Klass.__div__ = ... # SWIG magic

so that it fails.