Fix regression in nonzero guess logic in PETScKrylovSolver
The regression introduced in 5a10fbcc4b9977362a749caf02507173cdc70db7
causes that if empty vector is provided than zero guess is used in
subsequent solves regardless of what user set before the first solve.

Keep it on PETSc to set the default (which is false, i.e. zero guess).