Make the SWIG docstring generator more robust
The docstring generator will check if all files that should be
generated are actually generated. The problem came about since
doxygen 1.8.13 has a regression and segfaults when parsing
dolfin's C++ code.

With this fix dolfin can be built regardless of doxygen version,
but the Sphinx documentation will not work with doxygen 1.8.13